reset: reset-scmi: Match scmi device by both name and protocol id
[ Upstream commit 34ce3c5e691206d388fcc61f3da2ce8fe8a1907a ]

The scmi bus now has support to match the driver with devices not only
based on their protocol id but also based on their device name if one is
available. This was added to cater the need to support multiple devices
and drivers for the same protocol.

Let us add the name "reset" to scmi_device_id table in the driver so
that in matches only with device with the same name and protocol id
SCMI_PROTOCOL_RESET.

Acked-by: Philipp Zabel <p.zabel@pengutronix.de>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>
Change-Id: Ib52651b9305c2541b1f367f8b0cda0503391b83d