firmware: arm_scmi: Change xfer_block allocation scheme
There are multiple issues with the per message virtio transport specific
additional space management:

a) At this time, it has a serious bug. Each element of the 'xfer_block'
array increased by size of transport-specific headers, but address
scheme was not changed. So second element of this array points to the
'extra_data' member of the first element and so on, it is somehow works
only because in the workload only the first element of this array is
used.
b) Current approach uses too many virtio queue descriptors.
c) RX queue will work with messages with the different header sizes
(notifications and delayed responses), it will be necessary to correctly
set SCMI receive buffer pointer in each received message according to
the message type.

This commit fixes all of this:
a) Now there is an array of pointers to transfer blocks, each transfer
block allocated separately.
b) Each TX channel message uses 2 vqueue descriptors, and only one descriptor
for the RX channel.
c) Issue addressed.

Bug: 152907237
Test: build, iio test
Change-Id: I390f6dd60fe563c01e0c15a0e5ef14247eaf0e8d