firmware: arm_scmi: Move scmi protocols registration into the driver
[ Upstream commit 1eaf18e35a783a007ef03e09f1dfc3de81eace7c ]

In preparation to enable building SCMI as a single module, let us move
the SCMI protocol registration call into the driver. This enables us
to also add unregistration of the SCMI protocols.

The main reason for this is to keep it simple instead of maintaining
it as separate modules and dealing with all possible initcall races
and deferred probe handling. We can move it as separate modules if
needed in future.

Link: https://lore.kernel.org/r/20200907195046.56615-4-sudeep.holla@arm.com
Tested-by: Cristian Marussi <cristian.marussi@arm.com>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>
Change-Id: Ie65514e3dd4ca6e7311b4e360e7a41f0ed706404